requirements.txt
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.1.1
python-telegram-bot==22.4